Style category-filter
